support exponential backoff when tailing files (#17)
This commit extends the API of watch.NewPollingFileWatcher to accept
options for how frequently files should be polled. By default, if the
zero value is given, the previous value of polling every 250ms will be
used.

The customizable polling frequency supports exponential backoff; the
polling frequency will be doubled every time polling results in no
changes up to hitting MaxPollFrequency. The polling frequency is reset
to MinPollFrequency when a file change is detected.